% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_each_page.R
\name{fetch_each_page}
\alias{fetch_each_page}
\title{Fetches multiple pages of data from an API}
\usage{
fetch_each_page(
  .url,
  .instance = "live",
  .data = data.frame(),
  .max_return,
  .return,
  .initial_skip = 0
)
}
\arguments{
\item{.url}{the URL endpoint that we are fetching data from}

\item{.instance}{"live" (the default) or "test"}

\item{.data}{the data frame being used to gather all of the returned hrefs}

\item{.max_return}{a positive integer that limits the number of returned links}

\item{.return}{"newest" (the default) or "oldest" which is used when .max_return is provided}

\item{.initial_skip}{initial number of records to omit. Not for use by user}
}
\value{
a data frame with API hrefs ready for fetch_href()
}
\description{
Function that recursively calls itself to fetch 40 results at a time. This function is used by the `fetch_filter()` function, and users should therefore NOT need to directly make use of this function.
}
\seealso{
[fetch_href()]
}
