% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_filter.R
\name{fetch_filter}
\alias{fetch_filter}
\title{Fetch initial data from an API using a filter}
\usage{
fetch_filter(
  .api,
  .filter = "",
  .instance = "live",
  .return = "newest",
  .max_return = Inf
)
}
\arguments{
\item{.api}{name of API requesting information from, such as "Candidate", "AnalyticsResult"}

\item{.filter}{vector of characters that specify the filters being applied. Note that if you are filtering on dates and using 'ge' and 'le' , then here 'ge' means 'greater than or equal to', but 'le' means 'strictly less than' (and not 'less than or equal to').}

\item{.instance}{"live" (the default) or "test"}

\item{.return}{"newest" (the default) or "oldest", which is used when .max_return is provided}

\item{.max_return}{a positive integer that limits the number of returned links. Default value is 'Inf'}
}
\description{
Function to fetch the initial filtered data from the Surpass API. For several contextual examples of the use of this function, see the code chunks in HTML version of the package vignette
}
\seealso{
[fetch_href()]
}
