% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_href.R
\name{fetch_href}
\alias{fetch_href}
\title{Fetch data using provided href URL links}
\usage{
fetch_href(
  .data,
  .col = "href",
  .instance = "live",
  .fails = "keep",
  .append = ""
)
}
\arguments{
\item{.data}{the data frame that contains hrefs}

\item{.col}{"href" (the default) or the column name to extract the hrefs from.}

\item{.instance}{"live" (the default) or "test"}

\item{.fails}{"keep" (the default) or "drop". This option is used to monitor if any hrefs fail. The returned data frame populates a column called fetch.href with either "success" or "fail".}

\item{.append}{name of linked API to append to href, such as "/TestForms", "/ItemResponse", "?showItemResponse=true", etc. See Vignette API summary for append text in bold.}
}
\value{
a data frame with many columns of data fetched from an API
}
\description{
Function to fetch data using the href's of interest from the declared column of the provided data frame. For several contextual examples of the use of this function, see the code chunks in the HTML version of the package vignette.
}
\seealso{
[fetch_filter()] for use before this function, [flatten_fetched()] for use after this after
}
