% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_fetched.R
\name{flatten_fetched}
\alias{flatten_fetched}
\title{Flatten data frames that contain (nested) lists}
\usage{
flatten_fetched(.data, .depth = 1)
}
\arguments{
\item{.data}{the data frame that contains lists of lists}

\item{.depth}{positive integer to determine how many sub-lists of lists to unnest. Default value is 1.}
}
\value{
a data frame with more columns than previously, from unnesting lists. Rows where fetch.href was "fail" are not dropped.
}
\description{
Function to process the various lists that are within lists, to the specified depth of nested lists. For several contextual examples of the use of this function, see the code chunks in the HTML version of the package vignette.
}
\details{
Repeated application of the `flatten_fetched()` function is equivalent to changing the value of `.depth`. For example, ` |> flatten_fetched() |> flatten_fetched() |> flatten_fetched()` is the same as ` |> flatten_fetched(.depth = 3)`
}
