% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_interval.R
\name{format_interval}
\alias{format_interval}
\title{Format a time duration into seconds, minutes or HH:MM:SS}
\usage{
format_interval(.start, .end, .format = "s")
}
\arguments{
\item{.start}{a date-time in format YYYY-MM-DDTHH:MM:SS.SSS}

\item{.end}{a date-time in format YYYY-MM-DDTHH:MM:SS.SSS}

\item{.format}{"s" (the default) for seconds, or "m" for minutes, or "hms" for HH:MM:SS}
}
\value{
either a numerical value or a character string, depending on the chosen .format parameter
}
\description{
Function to take two input date-times and output the duration between them in HH:MM:SS format.
}
\examples{
\dontrun{
# calculate the duration of the assessment to the nearest minute
dplyr::mutate(timeTaken = format_interval(.start = startedDate,
                                          .end = submittedDate,
                                          .format = "m"))
}

}
