% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthen_fetched.R
\name{lengthen_fetched}
\alias{lengthen_fetched}
\title{Lengthen data frame that has multiple, similar column names}
\usage{
lengthen_fetched(.data, .na_rm = "")
}
\arguments{
\item{.data}{data frame that contains multiple similar columns, differing by numerical insertions and suffixes}

\item{.na_rm}{name of column within which all rows containing NA values are removed.}
}
\value{
a data frame with fewer columns and more rows. Rows where fetch.href was "fail" are removed (as they contain no data).
}
\description{
Function to reshape a data frame after using `flatten_fetched()` when multiple column names differ by numerical insertions and suffixes. This functions returns a longer data frame than that provided, which includes the column `lengthen_id` that records the insertions and suffixes that used to be part of the column names that have been renamed with an underscore character at their end. #' For several contextual examples of the use of this function, see the code chunks in the HTML version of the package vignette.
}
\seealso{
[lengthen_itemResponse()]
}
