% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthen_itemResponse.R
\name{lengthen_itemResponse}
\alias{lengthen_itemResponse}
\title{Lengthen data frame that contains item response data}
\usage{
lengthen_itemResponse(.data)
}
\arguments{
\item{.data}{data frame that contains multiple blocks of columns, one block for each question type}
}
\value{
a data frame with fewer columns and more rows
}
\description{
Function to pivot longer on the different question types that are returned when using the `/itemResponse` API. This function returns a longer data frame than that provided, which includes the column `item_type` that records the type of item that the subsequent columns contain data upon.
}
\examples{
\dontrun{

fetch_filter(.api = "AnalyticsResult",
             .filter = "",
             .max_return = 10) |>
  # extract the appropriate itemResponse hrefs by appending to the 'href' column's contents
  fetch_href(.append = "/ItemResponse") |>
  # fetch the data from the column 'href' using the "test" instance
  fetch_href() |>
  # deal with lists within columns, to the default depth of 1
  flatten_fetched() |>
  # reshape details for each item from wide to long, by question type
  lengthen_itemResponse()
}

}
\seealso{
[lengthen_fetched()]
}
