% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_data_cut.R
\name{monthly_data_cut}
\alias{monthly_data_cut}
\title{Save or Read in monthly data cuts from a nominated API}
\usage{
monthly_data_cut(.api, .action, .folder, .start, .end)
}
\arguments{
\item{.api}{name of API to work with. One of "Result", "AnalyticsResult" or SummaryResult"}

\item{.action}{"save" to fetch from AnalyticsResult API (taking roughly 1h30m per month) or "read" to load in encrypted data cuts}

\item{.folder}{directory location for the data cuts e.g "api_data_cuts/encrypted"}

\item{.start}{date of first month, written in "YYYY_MM" format}

\item{.end}{date of last month, written in "YYYY_MM" format}
}
\value{
depending on the .action parameter, this either saves an encrypted .rds file with no return; or reads in and returns a row-binded data frame of multiple months' data.
}
\description{
Function to either fetch monthly data from the nominated API, flatten, remove candidate Scottish Candidate Numbers (SCN), encrypt and save in stated folder; or to read in previously encrypted data cuts for analysis.
}
\examples{
\dontrun{

# fetch the AnalyticsResult API data for month of April 2025 and store the resulting encrypted file in a previously created folder called "temp"

monthly_data_cut(.api = "AnalyticsResult",
                 .action = "save",
                 .folder = "temp",
                 .start= "2025_04",
                 .end = "2025_04")

# read in the four months' encrypted AnalyticsResult data cuts from January 2025 to April 2025 inclusive, and store the single resulting data frame in `df`

df = monthly_data_cut(.api = "AnalyticsResult",
                      .action = "read",
                      .folder = "temp",
                      .start= "2025_01",
                      .end = "2025_04")
}
}
